@@ -49,22 +49,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.2.0",
-    "@angular/cdk": "^13.2.0",
-    "@angular/common": "^13.2.0",
-    "@angular/compiler": "^13.2.0",
-    "@angular/core": "^13.2.0",
-    "@angular/forms": "^13.2.0",
-    "@angular/material": "^13.2.0",
-    "@angular/platform-browser": "^13.2.0",
-    "@angular/platform-browser-dynamic": "^13.2.0",
-    "@angular/platform-server": "^13.2.0",
-    "@angular/router": "^13.2.0",
+    "@angular/animations": "^14.0.2",
+    "@angular/cdk": "^14.0.2",
+    "@angular/common": "^14.0.2",
+    "@angular/compiler": "^14.0.2",
+    "@angular/core": "^14.0.2",
+    "@angular/forms": "^14.0.2",
+    "@angular/material": "^14.0.2",
+    "@angular/platform-browser": "^14.0.2",
+    "@angular/platform-browser-dynamic": "^14.0.2",
+    "@angular/platform-server": "^14.0.2",
+    "@angular/router": "^14.0.2",
     "@fortawesome/angular-fontawesome": "^0.10.0",
     "@fortawesome/fontawesome-svg-core": "^6.1.0",
     "@fortawesome/free-regular-svg-icons": "^6.0.0",
     "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@nguniversal/express-engine": "^13.1.0",
+    "@nguniversal/express-engine": "^14.0.1",
     "ag-grid-angular": "^26.2.0",
     "ag-grid-community": "^26.2.0",
     "argparse": "^2.0.1",
@@ -75,14 +75,14 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.2.1",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "^13.2.0",
-    "@angular/compiler-cli": "^13.2.0",
+    "@angular-devkit/build-angular": "^14.0.2",
+    "@angular-eslint/builder": "14.0.0-alpha.3",
+    "@angular-eslint/eslint-plugin": "14.0.0-alpha.3",
+    "@angular-eslint/eslint-plugin-template": "14.0.0-alpha.3",
+    "@angular-eslint/schematics": "14.0.0-alpha.3",
+    "@angular-eslint/template-parser": "14.0.0-alpha.3",
+    "@angular/cli": "^14.0.2",
+    "@angular/compiler-cli": "^14.0.2",
     "@nightwatch/schematics": "^0.1.9",
     "@types/argparse": "^2.0.2",
     "@types/chart.js": "^2.9.34",
