@@ -13,7 +13,7 @@
 */
 import { DOCUMENT } from "@angular/common";
 import { Component, type OnInit, ViewChild, Inject } from "@angular/core";
-import { FormControl } from "@angular/forms";
+import { UntypedFormControl } from "@angular/forms";
 import type { MatStepper } from "@angular/material/stepper";
 import { Router } from "@angular/router";
 
@@ -72,25 +72,25 @@ export class NewDeliveryServiceComponent implements OnInit {
 	public deliveryService: DeliveryService = {...defaultDeliveryService};
 
 	/** Allows the user to set 'active' */
-	public activeImmediately = new FormControl();
+	public activeImmediately = new UntypedFormControl();
 	/** Allows the user to set 'bypass*' fields */
-	public bypassLoc = new FormControl("");
+	public bypassLoc = new UntypedFormControl("");
 	/** Allows the user to set 'cdn'/'cdnName' */
-	public cdnObject = new FormControl("");
+	public cdnObject = new UntypedFormControl("");
 	/** Allows the user to set the 'longDesc' */
-	public description = new FormControl("");
+	public description = new UntypedFormControl("");
 	/** Allows the user to set 'ipv6Enabled' */
-	public disableIPv6 = new FormControl();
+	public disableIPv6 = new UntypedFormControl();
 	/** Allows the user to set the 'displayName'/'xml_id' */
-	public displayName = new FormControl("");
+	public displayName = new UntypedFormControl("");
 	/** Allows the user to set 'type'/'typeId' */
-	public dsType = new FormControl();
+	public dsType = new UntypedFormControl();
 	/** Allows the user to set 'infoUrl' */
-	public infoURL = new FormControl("");
+	public infoURL = new UntypedFormControl("");
 	/** Allows the user to set 'originFqdn' */
-	public originURL = new FormControl("");
+	public originURL = new UntypedFormControl("");
 	/** Allows the user to set 'protocol' */
-	public protocol = new FormControl();
+	public protocol = new UntypedFormControl();
 
 	/** Need This to be a property for template access. */
 	public readonly protocolToString = protocolToString;
